requirements.txt
requests>=2.32.0
urllib3>=2.2.0
python-dotenv>=1.0.0
shodan>=1.31.0
python-whois>=0.9.4
dnspython>=2.6.0
